add error handling in shell script

+++ docker/rundocker.sh
@@ -3,6 +3,37 @@
 IMAGE_NAME="nenuoj-judger"
 CONTAINER_NAME="nenuoj-judger"
 
+if [ ! -n "$CONFIG_FILE_PATH" ]; then
+  echo "Please set environment variable CONFIG_FILE_PATH, such as /judger/config.ini"
+  exit 1
+fi
+
+if [ ! -n "$TEST_FILES_PATH" ]; then
+  echo "Please set environment variable TEST_FILES_PATH, such as /judger/test_files"
+  exit 1
+fi
+
+if [ ! -n "$SPJ_FILES_PATH" ]; then
+  echo "Please set environment variable SPJ_FILES_PATH, such as /judger/spj_files"
+  exit 1
+fi
+
+if [ ! -n "$LOGS_PATH" ]; then
+  echo "Please set environment variable LOGS_PATH, such as /judger/logs"
+  exit 1
+fi
+
+if [ ! -n "$HOSTPORT" ]; then
+  echo "Please set environment variable HOSTPORT, such as 27015"
+  exit 1
+fi
+
+if [ ! -n "$CONTAINERPORT" ]; then
+  echo "Please set environment variable CONTAINERPORT, such as 27015"
+  exit 1
+fi
+
+
 docker container rm -f $CONTAINER_NAME 1>/dev/null 2>&1
 
 docker run \
